adds remote-future-call
The tests are currenly commented out because they are giving
false-negatives.  I've manually tested remote-future-call and it works,
but for the test cases don't behave as I'd expect them to.  They
distributed tasks always seem to run on the same instance (the local
instance) even when specified not to.  When the members are in different
JVMs it works fine (as I have tested manually), but for some reason
Hazelcast seems to treat all the local (respective to the JVM/process)
 members as the same.

To see remote-future-call in action I documented my manual test process
in this screencast:

http://screencast.com/t/B40gEVJP68r